Add allocator_api2, AnyMemory
