+++ hil-test/Cargo.toml
@@ -215,6 +215,7 @@ harness           = false
 required-features = ["esp-wifi", "esp-alloc"]
 
 [dependencies]
+allocator-api2    = { version = "0.2.0", default-features = false, features = ["alloc"] }
 cfg-if             = "1.0.0"
 critical-section   = "1.1.3"
 defmt              = "0.3.8"
